Cleaned up ChatServlet and moved text and emoji parsing into different folder
